异常对应一致性都取自汇总表

+++ obd-core/src/main/java/com/wkhmedical/po/DeviceTime.java
@@ -59,7 +59,7 @@ public class DeviceTime implements Serializable {
 
 	private Long cks;
 
-	private Integer exprt;
+	private Long exprt;
 
 	private Long rps;
